separate pruneNonCriticalCss func to clarify executed in browser sandbox
